# src/shared/swagger/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Go Fsat Delivery",
    "description": "My Api node for delivery services",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000"
    }
  ],
  "paths": {
...
    "/api/features/": {
      "get": {
        "summary": "Get all Feature .",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Create new Feature .",
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "feature": {
                  "code": "getting_started",
                  "title": "Getting Started",
                  "type": "basic",
                  "subtitle": "getting started",
                  "icon": "x-icon",
                  "link": "/",
                  "order": 1,
                  "status": "active",
                  "parentFeature": "FeatureId"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/features/{id}": {
      "get": {
        "summary": "Get one Feature .",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "summary": "Update one Feature .",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "feature": {
                  "code": "getting_started",
                  "title": "Getting Started",
                  "type": "basic",
                  "subtitle": "getting started",
                  "icon": "x-icon",
                  "link": "/",
                  "order": 1,
                  "status": "active",
                  "parentFeature": "FeatureId"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete one Feature .",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/menus/features": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}